Allow PerByte to take smart pointers to the Reader instance.
This reflects the fact that PerByte consumes all of the bytes anyway, leaving
the Reader in an unusable state afterwards. So it might as well own the Reader
by storing it in a unique_ptr that is passed from the constructor, which
simplifies ownership concerns and lifetime management for users.