Overhauled the Reader::load so that it returns validity.
This enables a 'while (reader.load())' loop, where the buffer getters
can only be used inside the loop body, i.e., when 'load()' returns true.
This is more intuitive than a 'do/while' loop where a false return value
represents the last iteration and the caller needs to remember to break
out of the loop at the end of the iteration.

The new contract is slightly more difficult to implement for subclasses
that return bytes and the EOF in the same call, as the fact that the EOF
was encountered only takes effect in the next 'load()' call. But on the
whole, the performance penalty is minor and the usability is improved.