Parameterize the PerByte pointer type to accept raw/unique/shared pointers.
This gives the caller more control over how they wish to manage the lifetime of
the Reader passed to PerByte constructors. For example, if the input stream
still needs to be used after some PerByte processing, the caller can pass in a
raw pointer to avoid deletion after the PerByte instance is destroyed.

Also did the same for the IstreamReader and OstreamWriter classes, to allow
the calling context to maintain control over the lifetime of the stream.