Improvements to the integer type handling of gzip reader.
We limit the buffer size to fit in both an int and an unsigned, so that we
don't get an error when gzread() cannot fit the return value into an int.

We check for a negative return value from gzread() in case of error. We also
check for Z_OK upon reaching EOF to ensure that the read was completed okay.