Tightened up some of the docs for the interfaces.
In particular, we consider the case where some poor sucker is calling extract()
with n=0 after the byte source has been exhausted. On one hand, it is silly to
call get()/advance() at all if we know no more bytes are available; on the other
hand, extract() is meant to be equivalent to a loop of 'n' get()/advance()
calls, and if n=0, one could argue that extract() should be a safe no-op.

As it is currently written, the extract() code is only a safe no-op in this
case if the abstract Reader::load() can be safely called after it has returned
false once. This is too much of an imposition for Reader implementers, so
we're not doing that. Alternatively, we could add a code path to extract() to
explicitly check for n=0 after the input is exhausted, but this would be such
an unusual case that I don't want to dedicate any special casing to it.

So the simplest solution is to just admit that this scenario is silly, and
forbid extract() from being called in such silly cases.